+++ R/metadata.R
@@ -68,9 +68,10 @@ get_colour_scale <- function(iso3 = "default") {
 #' @examples
 #' get_metadata()
 get_metadata <- function() {
-  data <- naomi_read_csv(system_file("metadata", "metadata.csv"))
+  data <- naomi_read_csv(system_file("metadata", "metadata.csv"),
+                         na = character()) ## We want empty cells to be ""
   data$name <- traduire::translator()$replace(data$name)
-  
+
   ## TODO: refactor these into one location (issue #145)
   other_data <- get_meta_indicator()